test(test_hello): move integration test into tests folder
